Add pass-thru is_cloud for normalize_user
The action for Get Comments never relays the state of the endpoint as
on-premise or cloud, but requires this state to return successfully.

By using the same process as is present in other actions, passing of
the client is_cloud boolean through the normalize_comment function
and on to the normalize_user function with a default value of False.